+++ CHANGELOG.md
@@ -4,11 +4,19 @@
 
 ## Unreleased
 
-[compare changes](https://github.com/senekor/ec2hx/compare/v1.3.0...HEAD)
+[compare changes](https://github.com/senekor/ec2hx/compare/v1.4.0...HEAD)
 
 ### Breaking
 
 ### Changed
+
+### Added
+
+### Fixed
+
+## 1.4.0 - 2025-02-08
+
+[compare changes](https://github.com/senekor/ec2hx/compare/v1.3.0...v1.4.0)
 
 ### Added